feat: 로그아웃 구현

+++ src/main/kotlin/thatline/localup/property/TokenProperty.kt
@@ -10,6 +10,6 @@ data class TokenProperty(
 ) {
     data class AccessToken(
         var name: String,
-        val expirationSeconds: Int,
+        val expirationSeconds: Long,
     )
 }
